first attempt at coalescing 'x' and 'view'
When Adrien forked Waylon he took a copy of the code as-is and started
hacking in 'x'. This had a lot of awesome benefits, such as implementing
a proper JSON API and using Javascript to do all the display work.

This commit is a first attempt at coalescing the pre-existing 'view'
code and the new 'x' code. As of this commit, several things are still a
WIP or broken:
  - Servers that couldn't be contacted should display an error above the
    roll-up stats
  - Jobs that are unknown/invalid shouldn't be in the radiator; warnings
    should be displayed above the roll-up stats
  - Job investigation functionality needs to be re-implemented; see
    https://github.com/rji/waylon/blob/db1ed06/views/data.erb
  - The "loading" div might no longer be needed, since the view is
    pre-populated with a list of jobs
  - Some sorting wonkiness: it looks like jobs aren't being sorted in
    alphabetical order. Not sure if this is intentional.