refactor 'investigate' endpoint
Prior to this commit, the '*/investigate' endpoint was a piece of crap.
When a friendly server name was passed-in, it parsed the config looking
for a server that matched a regex so that the entire URL didn't have to
appear in the user's request.

This commit makes use of the manadic bits that Adrien implemented to do
accomplish the same thing much more efficiently.