Correctly set investigate menu element
Previously, investigate menu views could be in existence but would not
be bound to the DOM, and events on menu items could propagate in very
strange ways. As a result of this, setting a status on a build would set
it on ALL. BUILDS. This resolves this by setting the element on the
vialkjfsdmore
specifically.