implement icons for showing the job status
Prior to this commit, Waylon's primary means for communicating a job's
state is with colors. However, that makes Waylon unusable for any team
members that might be colorblind.

This commit implements Glyphicons for showing the job status next to
the job stability (weather) icon. For example: in addition to a table
row being green for a successful job, a checkmark is now also displayed.

This resolves gh-43.