refactor jobinfo table using bootstrap grid system
Prior to this commit, the radiator view used a table to display jobs.
While this was okay for the original prototype, we've added features and
started abusing the way tables are intended to be used for.

This commit refactors the table to use divs, and more specifically,
Bootstrap's grid system, for fun and profit.