fix display issues with stats roll-up
Prior to this commit, each of the stats roll-up descriptions/values were
h3 elements. When the viewport size was reduced, sometimes the text
would wrap and the job count would break from its label.

This commit uses the Bootstrap grid system to contain these in their own
columns, and ensures that the stats-rollup contents are responsive to
changes in viewport size.