don't display stats roll-up while loading
There was a small bug in the way stats-rollup was displayed, leftover
from the first iteration of Waylon (before things got all Ajaxy and
API-y). The elements for displaying the stats roll-up were still
included in the Ruby ERB template, causing them to be displayed even
when the "loading data" indicator was shown. In most cases, this was
almost immediately dismissed, but for any non-LAN connections, it became
quite obvious.

This commit removes the elements from the erb template now that they're
created by the Javascript side of the app.