(gh-48) correct bug in setting build description
When attempting to set a build description via the Waylon UI, the
description was never set. Watching the foreman logs, I noticed the
following exception being raised:

  JenkinsApi::Exceptions::NotFound - Requested component is not found
  on the Jenkins CI server.:

I inserted a quick `p prefix` in rest.rb and found that the job name
wasn't being escaped before the API call was being made.

Note that Jenkins requires spaces to be replaced by '%20', not '+',
meaning that CGI.escape would not work here.

This resolves gh-48.