(gh-51) fix jobs with no eta showing as unknown
Prior to this commit, jobs with no ETA, that is, Jenkins couldn't
calculate it or it was the first time the job ran, would appear as
unknown because jenkins_api_client was raising an exception that we
weren't handling.

This commit rescues the exception, and abuses Handlebars.js to check if
the eta value has been set to '-1'. If it has, display the text
'unknown'.

This resolves gh-51.