(gh-39) Radiator should use displayName
In job_details[], we have 'displayName' that returns a job's given
Display Name as set in the Jenkins UI, or if one isn't set, returns the
job's actual name. For anyone using the Display Name functionality of
Jenkins, this ensures their Waylon instance would have easy-to-read job
names.

This resolves gh-39.