trouble mode breaks nirvana mode
If trouble_threshold is set in waylon.yml, the background image fails to
appear due to trouble.js overriding the CSS for "body". This adds some
rudimentary checks to ensure we're not already in Nirvana mode.

This should probably be refactored for the check code to live in a
single place, and call out to nirvana.js or trouble.js when we reach
that state, instead of duplicating efforts all over.