Return server names and URLs for View endpoint
When querying the View endpoint, the server URLs were being returned but
the names were not. This meant that it was not possible to figure out
the correct API endpoints from the View endpoint. This commit updates
the endpoint to return a hash of "servername": "serverurl". Since this
endpoint is informational only no other code needs to change.