display build number next to job name
This commit displays the build number next to the job name. If you're
using the Build Name Setter plugin, you could easily append the short
Git SHA (i.e. ${GIT_REVISION,length=9}) to the build number to get a
better idea of which code is green and which code is red.